ai/requirements.txt
Pillow
numpy
torch
torchvision
opencv-python
